requirements.txt
PyQt6>=6.6.0
pyautogui>=0.9.54
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
pyperclip>=1.8.2
Pillow>=10.1.0
pyinstaller>=6.2.0

